move regex specific logic to FileDataAsset
